Migrate tests from jasmine to vitest; retire the e2e suite
Replaces jasmine with vitest as the test runner and removes the jasmine
dependency entirely. This is the first step toward a testing framework
built on the engine as a stable seam, rather than on browser-driven
golden-state comparisons.

Runner migration:
  - `npm test` / `npm run test-watch` now run vitest. Test files import
    describe/it/expect explicitly rather than relying on globals, so that
    there is no ambiguity about where the test API comes from.
  - InjectionHelper's spy wrapper moves to vi.spyOn. vitest does not
    export `Procedure` from its main entrypoint and vi.spyOn's overloads
    reject a generic `keyof T`, so the spy type is rebuilt via `infer`.
    Return-type checking is preserved: mockReturnValue is still
    validated against the spied method's return type.
  - 92 specs before, 92 tests after.

E2E suite removed:
  The suite asserted against full serialized-game-state JSON goldens,
  which broke on any unrelated state-shape change and produced diffs
  that were not actionable. It required Postgres, a server, and Chrome,
  was wired through deflake.sh (3 runs, pass if any one passes), and was
  workflow_dispatch-only. It will be replaced by a small set of
  Playwright specs covering client/server/engine wiring only.

  Note for that work: src/e2e/util/game_data.ts contained reusable
  database fixture logic (setUpGameEnvironment, initializeUsers) that is
  worth resurrecting from history rather than reinventing.

Prober preserved:
  The prober monitors production every 5 minutes and imported from
  src/e2e/util/, so it would have broken when that directory was
  deleted. Its web driver moves to src/testing/webdriver.ts, trimmed to
  the methods it actually uses; the retained code is unchanged.

  setTestTimeout is replaced by config. jasmine's
  DEFAULT_TIMEOUT_INTERVAL covered specs and hooks with one knob, but
  vitest splits them and the driver is built in a beforeAll, so
  vitest.prober.config.mts raises both testTimeout and hookTimeout.

  Verified locally that the driver launches, navigates, and tears down
  against production. The authenticated /login-as/ path was not
  exercised (no LOGIN_KEY available locally); that code is unchanged.
