--- .github/workflows/e2e.yml
@@ -1,65 +1,0 @@
-name: E2E Tests
-
-on:
-  workflow_dispatch:
-
-jobs:
-  run-e2e:
-    name: Run E2E tests
-    runs-on: ubuntu-latest
-
-    permissions:
-      contents: read
-      checks: write
-
-    services:
-      postgres:
-        image: postgres:latest
-        env:
-          POSTGRES_DB: postgres
-          POSTGRES_PASSWORD: postgres
-          POSTGRES_USER: postgres
-        options: >-
-          --health-cmd pg_isready
-          --health-interval 10s
-          --health-timeout 5s
-          --health-retries 5
-        ports:
-          - 5432:5432
-
-    steps:
-      - name: Check out Git repository
-        uses: actions/checkout@v4
-
-      - run: sudo apt-get update
-      - run: sudo apt-get install unzip
-
-      - name: Set up Node.js
-        uses: actions/setup-node@v4
-        with:
-          node-version: 25
-
-      - name: Install Node.js dependencies
-        run: npm ci
-
-      - name: Set up Java
-        uses: actions/setup-java@v4
-        with:
-          java-version: "21"
-          distribution: "temurin"
-
-      - name: Build
-        run: npm run build
-
-      - name: Run migrations
-        run: npm run migrate-testdb-direct
-        env:
-          POSTGRES_URL: "postgres://postgres:postgres@localhost:${{ job.services.postgres.ports[5432] }}/postgres"
-          NODE_ENV: "test"
-
-      - name: Run E2E tests
-        run: ./src/scripts/deflake.sh "npm run e2e-test"
-        env:
-          POSTGRES_URL: "postgres://postgres:postgres@localhost:${{ job.services.postgres.ports[5432] }}/postgres"
-          NODE_ENV: "test"
-          PORT: 3001
