+++ package.json
@@ -17,13 +17,9 @@
     "migrate": "node bin/scripts/migrations up",
     "migrate-testdb": "export $(cat .testenv.sh | xargs) && npm run migrate-testdb-direct",
     "migrate-testdb-direct": "node bin/scripts/migrations up",
-    "test": "ts-node node_modules/jasmine/bin/jasmine --config=spec/support/jasmine.mjs",
-    "test-watch": "nodemon -V --ext ts,tsx --watch src/ --exec 'npm run test'",
-    "e2e": "export $(cat .testenv.sh | xargs) && npm run e2e-test",
-    "e2e-test": "ts-node node_modules/jasmine/bin/jasmine --config=spec/support/e2e.mjs",
-    "e2e-write": "WRITE=true npm run e2e",
-    "e2e-watch": "nodemon -V --ext ts,tsx --watch src/ --exec 'npm run e2e'",
-    "prober": "ts-node node_modules/jasmine/bin/jasmine --config=spec/support/prober.mjs",
+    "test": "vitest run",
+    "test-watch": "vitest",
+    "prober": "vitest run --config vitest.prober.config.mts",
     "knip": "knip"
   },
   "files": [
@@ -73,7 +69,6 @@
     "@types/cookie-parser": "^1.4.7",
     "@types/express": "^5.0.1",
     "@types/express-session": "^1.18.0",
-    "@types/jasmine": "^5.1.5",
     "@types/node": "^22.14.1",
     "@types/react": "^18.3.20",
     "@types/react-dom": "^18.3.6",
@@ -84,14 +79,14 @@
     "eslint-config-prettier": "^10.1.2",
     "eslint-plugin-react": "^7.37.5",
     "globals": "^15.14.0",
-    "jasmine": "^5.5.0",
     "knip": "^5.50.5",
     "nodemon": "^3.1.9",
     "prettier": "3.4.2",
     "selenium-webdriver": "^4.33.0",
     "ts-node": "^10.9.2",
     "typescript": "^5.8.3",
-    "typescript-eslint": "^8.30.1"
+    "typescript-eslint": "^8.30.1",
+    "vitest": "^4.1.10"
   },
   "author": "YourDeveloperFriend",
   "license": "ISC",
