--- spec/support/jasmine.mjs
@@ -1,15 +1,0 @@
-export default {
-  spec_dir: "",
-  spec_files: [
-    "src/**/*_test.ts",
-    "src/**/*_test.tsx",
-    "!src/e2e/*",
-    "!src/prober/*",
-  ],
-  jsLoader: "require",
-  env: {
-    stopSpecOnExpectationFailure: false,
-    random: true,
-    forbidDuplicateNames: true,
-  },
-};
