--- src/e2e/util/server.ts
@@ -1,10 +1,0 @@
-import { runApp } from "../../server/server";
-import { log } from "../../utils/functions";
-
-export function setUpServer(): void {
-  beforeAll(async function turnUpServer() {
-    log("start server turnup");
-    await runApp();
-    log("end server turnup");
-  });
-}
