--- src/e2e/util/timeout.ts
@@ -1,12 +1,0 @@
-export function setTestTimeout(timeout: number): void {
-  let originalTimeout: number;
-
-  beforeAll(() => {
-    originalTimeout = jasmine.DEFAULT_TIMEOUT_INTERVAL;
-    jasmine.DEFAULT_TIMEOUT_INTERVAL = timeout;
-  });
-
-  afterAll(() => {
-    jasmine.DEFAULT_TIMEOUT_INTERVAL = originalTimeout;
-  });
-}
