+++ src/engine/game/money_manager_test.ts
@@ -1,3 +1,4 @@
+import { describe, expect, it } from "vitest";
 import { city, plain } from "../../testing/factory";
 import { InjectionHelper } from "../../testing/injection_helper";
 import { resettable } from "../../testing/resettable";
