+++ src/maps/grid_order_test.ts
@@ -1,3 +1,4 @@
+import { describe, expect, it } from "vitest";
 import { isOrdered } from "immutable";
 import { Grid } from "../engine/map/grid";
 import { MapRegistry } from "../maps/registry";
