+++ src/maps/tile_factory_test.ts
@@ -1,3 +1,4 @@
+import { describe, expect, it } from "vitest";
 import { PlayerColor } from "../engine/state/player";
 import { Direction, SimpleTileType } from "../engine/state/tile";
 import { startFrom } from "./tile_factory";
